+++ TinyCSV.Example/Program.cs
@@ -9,48 +9,46 @@ namespace TinyCSV.Example
         {
             //Write.
             //Create a empty csv writer.
-            CSVTableWriter csvTableWriter = new CSVTableWriter();
-            //Add headers.
-            csvTableWriter.AddHeader("Data1");
-            csvTableWriter.AddHeader("Data2");
-            csvTableWriter.AddHeader("Data3");
-            csvTableWriter.AddHeader("Data4");
-            csvTableWriter.AddHeader("Data5");
-            csvTableWriter.AddHeader("Data6");
-            //Add descriptions.
-            csvTableWriter.AddDescription("int");
-            csvTableWriter.AddDescription("Color");
-            csvTableWriter.AddDescription("int");
-            csvTableWriter.AddDescription("string");
-            csvTableWriter.AddDescription("string");
-            csvTableWriter.AddDescription("int[]");
-            //Add records.
-            var record1 = new CSVRecordWriter();
-            record1.AddCell("1");
-            record1.AddCell("#aaaaaa");
-            record1.AddCell("2");
-            record1.AddCell("normal string");
-            record1.AddCell("\"string with double quote");
-            record1.AddCell("1;2;3|4;5;6");
-            csvTableWriter.AddRecord(record1);
-            var record2 = new CSVRecordWriter();
-            record2.AddCell("2");
-            record2.AddCell("#bbbbbb");
-            record2.AddCell("4");
-            record2.AddCell("string with, comma");
-            record2.AddCell("string\" with\", comma and \"double quote");
-            record2.AddCell("7;8;9|10;11;12|7;7;7");
-            csvTableWriter.AddRecord(record2);
-            var record3 = new CSVRecordWriter();
-            record3.AddCell("3");
-            record3.AddCell("#ccc");
-            record3.AddCell("5");
-            record3.AddCell("string with \n \\n and \r\n \\r\\n");
-            record3.AddCell("string\" with \n \\n and \", comma and \"double quote and \r\n \\r\\n");
-            record3.AddCell("7;8;9|10;11;12|7;7;7");
-            csvTableWriter.AddRecord(record3);
+            CSVTableWriter csvTableWriter = new CSVTableWriter()
+                //Add headers.
+                .AddHeader("Data1")
+                .AddHeader("Data2")
+                .AddHeader("Data3")
+                .AddHeader("Data4")
+                .AddHeader("Data5")
+                .AddHeader("Data6")
+                //Add descriptions.
+                .AddDescription("int")
+                .AddDescription("Color")
+                .AddDescription("int")
+                .AddDescription("string")
+                .AddDescription("string")
+                .AddDescription("int[]")
+                //Add records.
+                .AddRecord(new CSVRecordWriter()
+                    .AddCell("1")
+                    .AddCell("#aaaaaa")
+                    .AddCell("2")
+                    .AddCell("normal string")
+                    .AddCell("\"string with double quote")
+                    .AddCell("1;2;3|4;5;6")
+                )
+                .AddRecord(new CSVRecordWriter()
+                    .AddCell("2")
+                    .AddCell("#bbbbbb")
+                    .AddCell("4")
+                    .AddCell("string with, comma")
+                    .AddCell("string\" with\", comma and \"double quote")
+                    .AddCell("7;8;9|10;11;12|7;7;7"))
+                .AddRecord(new CSVRecordWriter()
+                    .AddCell("3")
+                    .AddCell("#ccc")
+                    .AddCell("5")
+                    .AddCell("string with \n \\n and \r\n \\r\\n")
+                    .AddCell("string\" with \n \\n and \", comma and \"double quote and \r\n \\r\\n")
+                    .AddCell("7;8;9|10;11;12|7;7;7"));
             //Get csv form string.
-            string csv1 = csvTableWriter.ToString();
+            string csv1 = csvTableWriter.GetEncodeTable();
             Console.WriteLine("Write to csv form:\n" + csv1);
             using (StreamWriter streamWriter = new StreamWriter(Path.Combine(Environment.CurrentDirectory, "CSVExample1.csv")))
                 streamWriter.Write(csv1);
@@ -61,23 +59,25 @@ namespace TinyCSV.Example
             Console.WriteLine("Read from csv form:\n" + csvTableReader);
             
             //Write
-            //Create csv writer from csv reader.
-            CSVTableWriter csvTableWriter2 = new CSVTableWriter(csvTableReader);
-            csvTableWriter2.RemoveHeader(5);
-            csvTableWriter2.RemoveDescription(5);
-            csvTableWriter2.RemoveRecord(0);
+            //Create csv writer from csv reader, custom cell separator.
+            char cellSeparator = ';';
+            CSVTableWriter csvTableWriter2 = new CSVTableWriter(csvTableReader, cellSeparator)
+                //Remove some row and column.
+                .RemoveHeader(5)
+                .RemoveDescription(5)
+                .RemoveRecord(0);
             foreach (var record in csvTableWriter2.Records)
                 record.RemoveCell(5);
-            var record100 = new CSVRecordWriter();
-            record100.AddCell("100");
-            record100.AddCell("#cccccc");
-            record100.AddCell("5");
-            record100.AddCell("string with, comma");
-            record100.AddCell("\"string with\", comma and \"double quote");
-            csvTableWriter2.AddRecord(record100);
-            Console.WriteLine("Write to csv form:\n" + csvTableWriter2);
-            //Get csv form string.
-            string csv2 = csvTableWriter2.ToString();
+            //Add new record.
+            csvTableWriter2.AddRecord(new CSVRecordWriter(cellSeparator)
+                .AddCell("100")
+                .AddCell("#cccccc")
+                .AddCell("5")
+                .AddCell("string with; semicolon")
+                .AddCell("string\" with\"; semicolon and \"double quote"));
+            //Get csv form string, choose new line style.
+            string csv2 = csvTableWriter2.GetEncodeTable(NewLineStyle.NonUnix);
+            Console.WriteLine("Write to csv form:\n" + csv2);
             using (StreamWriter streamWriter = new StreamWriter(Path.Combine(Environment.CurrentDirectory, "CSVExample2.csv")))
                 streamWriter.Write(csv2);